ENH: add simplified expr syntax
