+++ altair/expr/__init__.py
@@ -1,0 +1,5 @@
+"""Tools for creating transform & filter expressions with a python syntax"""
+
+from .core import datum
+from .funcs import *
+from .consts import *
